backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiofiles==24.1.0
redis==5.1.1
httpx==0.27.2
supabase==2.9.0
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-community==0.3.7
pgvector==0.3.5
python-dotenv==1.0.1
email-validator>=2.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
slowapi>=0.1.9
bandit==1.7.10
